requirements.txt
fastapi==0.111.0
langchain==0.3.0
langchain-google-genai==2.0.0
python-dotenv==1.0.1
pydantic==2.8.2
uvicorn==0.30.1